ajustes mr

+++ src/main/java/br/com/livelo/orderflight/mappers/ReservationMapper.java
@@ -1,13 +1,5 @@
 package br.com.livelo.orderflight.mappers;
 
-import java.time.LocalDateTime;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.factory.Mappers;
-
 import br.com.livelo.orderflight.domain.dto.reservation.request.PartnerReservationDocument;
 import br.com.livelo.orderflight.domain.dto.reservation.request.PartnerReservationRequest;
 import br.com.livelo.orderflight.domain.dto.reservation.request.ReservationDocument;
@@ -18,6 +10,12 @@ import br.com.livelo.orderflight.domain.entity.OrderEntity;
 import br.com.livelo.orderflight.domain.entity.OrderItemEntity;
 import br.com.livelo.orderflight.domain.entity.OrderPriceEntity;
 import br.com.livelo.orderflight.domain.entity.OrderStatusEntity;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.factory.Mappers;
+
+import java.util.Set;
+import java.util.stream.Collectors;
 
 
 @Mapper(componentModel = "spring", uses = {ReservationItemMapper.class, ReservationPriceMapper.class})
@@ -31,9 +29,9 @@ public interface ReservationMapper {
     @Mapping(target = "channel", source = "channel")
     @Mapping(target = "transactionId", source = "transactionId")
     @Mapping(target = "customerIdentifier", source = "customerId")
-    @Mapping(target = "statusHistory",  expression = "java(Set.of(mapStatus(partnerReservationResponse)))")
-	@Mapping(target = "status", expression = "java(mapStatus(partnerReservationResponse))")
-	@Mapping(target = "price", expression = "java(mapPrice(partnerReservationResponse, listPrice))")
+    @Mapping(target = "statusHistory", expression = "java(Set.of(mapStatus(partnerReservationResponse)))")
+    @Mapping(target = "status", expression = "java(mapStatus(partnerReservationResponse))")
+    @Mapping(target = "price", expression = "java(mapPrice(partnerReservationResponse, listPrice))")
     @Mapping(target = "createDate", ignore = true)
     OrderEntity toOrderEntity(ReservationRequest reservationRequest, PartnerReservationResponse partnerReservationResponse, String transactionId, String customerId, String channel, String listPrice);
 
@@ -69,6 +67,6 @@ public interface ReservationMapper {
     PartnerReservationDocument toPartnerReservationDocument(ReservationDocument reservationDocument);
 
     @Mapping(target = "expirationTimer", source = "expirationTimer")
-    ReservationResponse toReservationResponse(OrderEntity orderEntity, Long expirationTimer);
+    ReservationResponse toReservationResponse(OrderEntity orderEntity, int expirationTimer);
 
 }
